Actually take stars and arrows when used (and make them don't disappear when they're at 0)
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@292 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/Inventory.cpp
@@ -484,7 +484,7 @@ void Inventory::takeItemSlot(Player* player, short slot, char inv, short amount)
 		if(player->inv->getItem(i)->pos == slot && player->inv->getItem(i)->inv == inv){
 			Item* item = player->inv->getItem(i);
 			item->amount-=amount;
-			if(item->amount == 0){
+			if(item->amount == 0 && !ISSTAR(item->id) && !ISARROW(item->id)){
 				InventoryPacket::moveItem(player, item->inv, item->pos, 0);
 				player->inv->deleteItem(i);
 			}
